doc: document connection types
