+++ desktop/src/visual/mod.rs
@@ -1,5 +1,6 @@
 pub mod blocks;
 pub mod canvas;
+pub mod connections;
 pub mod palette;
 pub mod serialization;
 pub mod translations;
